requirements.txt
theano==1.0.3
pymc3==3.5
jax==0.2.12
jaxlib==0.1.65
neural-tangents==0.3.6
pytorch==1.7.1
